requirements.txt
Flask
flask-cors
joblib
scikit-learn
requests
python-dotenv
